requirements.txt
SQLAlchemy
Flask
gunicorn
pymysql
python-dotenv
bcrypt
itsdangerous
requests
Flask-Mail
twilio
